refactor(all managers): move all manager class into controller folder
